requirements.txt
Flask==3.1.2
python-telegram-bot==20.7
python-dotenv==1.0.0
Werkzeug==3.1.4
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
SQLAlchemy==2.0.45
alembic==1.17.2
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
click==8.3.1
blinker==1.9.0
itsdangerous==2.2.0
greenlet==3.3.0
colorama==0.4.6
typing_extensions==4.15.0
qrcode==7.4.2
bakong-khqr==0.5.7
Pillow==10.1.0